choose to remove local or remote when building

+++ src/build.ps1
@@ -1,5 +1,7 @@
 param (
-  [String]$FileToOpen = ""
+  [String]$FileToOpen = "",
+  [int]$RemoveRemote = 0,
+  [int]$RemoveLocal = 0
 )
 
 Write-Output "========================="
@@ -24,10 +26,14 @@ clang notes.c md_renderer.c md4c/md4c.c helpers.c -o notes.exe -Wno-deprecated-d
 
 Write-Output "Building notes..."
 
-Remove-Item "../remote/*.html"
-Remove-Item "../remote/tags/*.html"
-Remove-Item "../local/*.html"
-Remove-Item "../local/tags/*.html"
+if ($RemoveLocal -ne 0) {
+  Remove-Item "../local/*.html"
+  Remove-Item "../local/tags/*.html"
+}
+if ($RemoveRemote -ne 0) {
+  Remove-Item "../remote/*.html"
+  Remove-Item "../remote/tags/*.html"
+}
 
 .\notes.exe $notes
